Rewritten the whole javascript backend
New parser using acorn with the extra nerdlang structure.
New compiler with the acorn output, the formatting is alot cleaner.
It is now a multi-file output, instead of the cramming everything into a single file, it mirrors the input directory with its corresponding c++ files.
It uses CMake, to make it easier to support new platforms and adding external libraries in the future (currently only tested on Ubuntu).
With this approach, it will be possible to use the generated output as a library. (It is already possible, just not documented).